+++ Sources/LogConsole/Properties/AssemblyInfo.cs
@@ -25,7 +25,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can use the default the Revision and 
 // Build Numbers by using the '*' as shown below:
 [assembly: AssemblyVersion ("0.16.*")]
-[assembly: AssemblyInformationalVersion("0.16 for KSP v1.3+")]
+[assembly: AssemblyInformationalVersion("0.16 for KSP v1.4")]
 [assembly: KSPAssembly("KSPDev_LogConsole", 0, 16)]
 [assembly: Guid ("4a5a3eab-4159-4375-8329-4f739fcb6226")]
 [assembly: AssemblyFlags (AssemblyNameFlags.PublicKey | AssemblyNameFlags.EnableJITcompileOptimizer)]