+++ Sources/Utils/ConfigUtils/StdPersistentGroups.cs
@@ -2,6 +2,8 @@
 // Author: igor.zavoychinskiy@gmail.com
 // This software is distributed under Public domain license.
 
+namespace KSPDev.ConfigUtils {
+
 /// <summary>Group names that have special meaning.</summary>
 /// <seealso cref="KSPDev.ConfigUtils.ConfigAccessor"/>
 /// <seealso cref="KSPDev.ConfigUtils.PersistentFieldAttribute"/>
@@ -12,6 +14,7 @@ public static class StdPersistentGroups {
   /// persistent fields with a specific save/load logic, then they need to have a group different
   /// from the default.
   /// </remarks>
+  /// <seealso cref="PersistentFieldAttribute"/>
   public const string Default = "";
 
   /// <summary>
@@ -22,6 +25,17 @@ public static class StdPersistentGroups {
   /// methods: <c>OnSave</c> and <c>OnLoad</c>. Dealing with them outside of this logic is highly
   /// discouraged.
   /// </remarks>
+  /// <seealso cref="PersistentFieldAttribute"/>
   public const string PartPersistant = "PartPersistant";
+
+  /// <summary>A public group for the feilds that needs to be loaded from a part config.</summary>
+  /// <remarks>
+  /// It's a very special group, never deal with it directly. The consumer code is only allowed to
+  /// use this group when defining the persistent fields via annotations.  
+  /// </remarks>
+  /// <seealso cref="PersistentFieldAttribute"/>
+  /// <seealso cref="ConfigAccessor.ReadPartConfig"/>
+  public const string PartConfigLoadGroup = "PartConfig";
 }
 
+}  // namespace