+++ Sources/Utils/GUIUtils/GuiWindow.cs
@@ -8,6 +8,12 @@ using UnityEngine;
 namespace KSPDev.GUIUtils {
 
 /// <summary>A utility class to deal with the GUI windows.</summary>
+/// <remarks>
+/// The stock window cancels dragging when the window layout is changed. It makes it useless when
+/// dealing with windows that can change their layout depending on the position. This method
+/// doesn't have this drawback. Moreover, it can tell if the window is being dragged, so that the
+/// code could postpone the layout update until the dragging is over. 
+/// </remarks>
 /// <example><code source="Examples/GUIUtils/GuiWindow-Examples.cs" region="GuiWindowDemo1"/></example>
 public static class GuiWindow {
   /// <summary>Latest mouse position to which the dragged window position has updated.</summary>
@@ -18,12 +24,6 @@ public static class GuiWindow {
   /// Makes the window movable. It's an improved version of the stock <c>GUI.DragWindow()</c>
   /// method.
   /// </summary>
-  /// <remarks>
-  /// The stock method cancels dragging when the window layout is changed. It makes it useless when
-  /// dealing with windows that can change their layout depending on the position. This method
-  /// doesn't have this drawback. Moreover, it can tell if the window is being dragged, so that the
-  /// code could postpone the layout update until the dragging is over. 
-  /// </remarks>
   /// <param name="windowRect">
   /// The window rectangle. It must be the same instance which is passed to the
   /// <c>GUILayout.Window</c> method.