+++ Sources/LogConsole/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.16 (December 14th, 2017):
-* [Enhancement] Migrate to `KSPDev Utils v0.30`.
+* [Change] KSP 1.4.0 compatibility.
+* [Change] Use KSPDev_Utils `KSPDev Utils v0.31`.
 
 # 0.15 (July 15th, 2017):
 * [Enhancement #14] Allow three fixed window size modes: top, bottom, or center.