+++ Sources/Utils/CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.31-pre:
+* [Chage] KSP 1.4.0 compatibility.
+* [Enhancement] Add `onBeforeTransition` event into: [ProcessingUtils.SimpleStateMachine](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_ProcessingUtils_SimpleStateMachine_1).
+* [Enhancement] Better print `Vector3` and `Quaternion` objects with better precission in [LogUtils.DebugEx](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_LogUtils_DebugEx.htm).
+* [Enhancement] Add optional paraemetrs to skip the specified number of frames [ProcessingUtils.AsyncCall](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_ProcessingUtils_AsyncCall.htm).
+* [Enhancement] Add methods to inject/withdraw part's menu events: [PartUtils.PartModuleUtils](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_PartUtils_PartModuleUtils.htm).
+* [Enhancement] Add a new method to align vessels via the attach nodes: [AlignTransforms.SnapAlignNodes](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ModelUtils_AlignTransforms_SnapAlignNodes.htm).
+* [Enhancement] Allow restricting the state machine handlers to init/shutdown sequence: [SimpleStateMachine.AddStateHandlers](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ProcessingUtils_SimpleStateMachine_1_AddStateHandlers.htm).
+* [Enhancement] Add a new method to load custom fields from the part's config: [ConfigAccessor.ReadPartConfig](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ConfigUtils_ConfigAccessor_ReadPartConfig.htm).
+* [Enhancement] Add a localization class to format a velocity value in a human friendly format: [`GUIUtils.VelocityType`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_VelocityType.htm).
+* [Enhancement] Add a utility class to create `PushButton` and `ToggleButton` controls: [`GUIUtils.KSPUtilsGUILayout`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_KSPUtilsGUILayout.htm).
+* [Enhancement] Add a serializable `ConfigNode` to deal with a portion of the config in a form of simple node: [`Types.PersistentConfigNode`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_Types_PersistentConfigNode.htm).
+* [Fix] Properly handle string paths in [`ConfigUtils.ConfigAccessor`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_ConfigUtils_ConfigAccessor.htm): treat an empty string as a root object reference.
+
 # 0.30 (December 14th, 2017):
 * [Enhancement] Add new method to align vessel via the nodes: [AlignTransforms.SnapAlignVessel](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ModelUtils_AlignTransforms_SnapAlignVessel.htm).
 * [Enhancement] Add new method to place a vessel: [AlignTransforms.PlaceVessel](http://ihsoft.github.io/KSPDev/Utils/html/M_KSPDev_ModelUtils_AlignTransforms_PlaceVessel.htm).